Fix snapshot cast import
